+++ lib/pvectl.rb
@@ -183,6 +183,10 @@ require_relative "pvectl/commands/config/set_cluster"
 require_relative "pvectl/commands/config/set_credentials"
 require_relative "pvectl/commands/config/view"
 require_relative "pvectl/commands/config/command"
+require_relative "pvectl/commands/cloudinit/regenerate"
+require_relative "pvectl/commands/cloudinit/pending"
+require_relative "pvectl/commands/cloudinit/dump"
+require_relative "pvectl/commands/cloudinit/command"
 require_relative "pvectl/commands/ping"
 require_relative "pvectl/commands/apt"
 
@@ -195,6 +199,9 @@ require_relative "pvectl/services/container_lifecycle"
 
 # Services - Snapshot
 require_relative "pvectl/services/snapshot"
+
+# Services - Cloudinit
+require_relative "pvectl/services/cloudinit"
 
 # Services - Backup
 require_relative "pvectl/services/backup"